feat(wasi): Adds working status updating
Updating status is now working for wasi. This has all been added as part
of the `wasi_runtime::handle` module, and so should be portable to
wascc when it is ready. It introduces a new `PodHandle` which is a higher
level (and possibly better?) abstraction for interacting with a running Pod.

I also removed the `status` method from the provider interface as it should
be done with handles and not by the kubelet calling the provider