fix(oci): Fixes clippy error around default client
Clippy was warning that we could replace the `new` method with a derived
Default. I checked reqwest and found that `Client::default()` is the
name as calling `Client::new()`. So I switched the `new` method to take
the config object and replaced the original `new` method with the derived
Default.