fix the release workflow
This commit fixes two things:

We want the release workflow to start when the tag is pushed rather than when it is published on github. That way, the release assets are ready and available when the release is published.

It turns out that you cannot have two conditional jobs with the same ID. Go figure. Instead, we now inject the release version into the environment and run the step if the reference is valid.

Signed-off-by: Matthew Fisher <matt.fisher@microsoft.com>