ref(handle): Moves handle to kubelet crate and makes it fully generic
This adds a new `Stop` trait that can be implemented per provider while still preserving
all the pod and container management goodness of the handles. Yes, there are
now 2 generic types on the handle structs, but after discussing it with @rylev
we are ok with this as a middleground for now. This should enable any provider to
take advantage of these handles.

Please note that I removed the test for the wasi runtime because with the new
implementation, we can't wait on things to complete right now. Also, the same
functionality is being covered in the e2e tests