fix(*): Fixes Rustls builds
We were missing some feature enablement on the reqwest dependency.
I also took the time to update us to the latest version of the `kube`
crate. This has been tested against a "real" (AKS in this instance)
Kubernetes cluster and works great. There are still problems with
running it with minikube, kind, and any other cluster that doesn't
have an actual FQDN. This is an upstream issue that we'll need to
tackle separately