chore(*): Bumps all deps to the latest version
This is just a general hygiene PR to update all dependencies possible
to their latest version. There were small code changes to update with
the new APIs from some libraries. I was not able to update `hyperx`
because our www-authenticate crate depends on a trait from the older
version.

I also can't update wasmtime to 0.20 yet due to wascc requiring 0.19.
We'll have to stick with it until we split out the crates or wascc upgrades