chore(kubelet): Removes unneeded stream wrapper
This is some follow up to comments in #480. I have one more to make once
I am on a Windows machine (#531). I also considered having the Krator `Manifest`
use a `WatchStream` wrapper, but that has several trait bounds that we don't need
so I chose not to use it.

This also contains some minor version bumps for things