fix(plugin): Avoids early termination of kubelet on startup
I discovered this bug when testing. If a CSI plugin fails to clean up
its socket file in the plugin directory, on restart kubelet will exit with
an error. This addresses that issue and does a little bit of minor upgrades
in the various project crates (because it was annoying me). Oh, and I
reenabled kubelet logging in the justfile

You can create a fake socket file in your plugins directory like so:

```
$ touch ~/.krustlet/plugins/foo.sock
```

With this PR, you should now just see an error message logged, but kubelet
will not exit