Support DockerHub OCI Authentication (#554)
* Allow duplicate token fields in OCI auth JSON

When requesting an authorization token from Docker Hub,
the JSON response contains both a 'token' and 'access_token'
field (with identical values). Since 'access_token' was used
as an alias field, this was causing the serde deserialize to
fail due to "duplicate field `token` at line 1 column X".

This could have been fixed by adding both fields to the struct,
marking both as optional. However, this method would have
meant that a successful deserialization using serde would not
necessarily indicate that a token was found in either field,
and would require additional logic in the caller.

To solve this, a custom deserialize implementation has been
added for RegistryToken, which allows for both fields. In the
case that both fields are present (i.e. Docker Hub), then the
'token' field takes precedence and 'access_token' is ignored.

Fixes #495.

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* fix cargo fmt check

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* do not make RegistryToken public

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* add comments to clarify integer tests

Signed-off-by: Josh Dolitsky <josh@dolit.ski>

* Serde untagged attempt

Co-authored-by: Josh Dolitsky <josh@dolit.ski>