oci-distribution: fix push to ghcr.io
The oci-distribution crate is currently pushing layers using the
"monolithic upload" API offered by distribution v2[1].

According to the official specification, the layer has to be uploaded
using this API:

```
PUT /v2/<name>/blobs/uploads/<uuid>?digest=<digest>
```

Prior to this commit, the code was building the destination url
using simple string interpolation.
That resulted in making a request like:

```
PUT /v2/<name>/blobs/uploads/<uuid>&digest=<digest>
```

Interestingly enough, this didn't seem to be a problem to the
distribution servers. Apparently the Go implementation is more
relaxed when it comes to parsing an incoming request.

The GitHub Package registry (ghcr.io), on the other hand rejects these
requests with a 404 response.

This commits fixes the problem by using the `Reqwest::Url` structure to
create the target url.

[1] https://github.com/distribution/distribution/blob/v2.7.1/docs/spec/api.md#monolithic-upload

Signed-off-by: Flavio Castelli <fcastelli@suse.com>