refactor(oci-distribution): use constructor request wrapper type
Introduce the `RequestBuilderWrapper` internal type that allows to
call to request building functions that can consume information stored
in the client.

The main goal of the wrapper type is that functions that are
implemented for it that alter the request builder wrapper are
composable.

When all the composing that requires some context from the client has
been done, `into_request_builder()` is called, handing a regular
`RequestBuilder` type that can be used to further add more information
to the request, and eventually, trigger it.

Signed-off-by: Rafael Fernández López <rfernandezlopez@suse.com>